requirements.txt
Django==1.5.2
Markdown==2.5.2
MySQL-python==1.2.3
Pillow==2.7.0
Pygments==2.0.1
python-memcached==1.53
wsgiref==0.1.2
